+++ src/functions/init/intro.js
@@ -5,11 +5,9 @@ const {
   AUTHOR,
   VERSION,
   BUILD,
-  RELEASE_CANDIDATE,
   IN_BETA,
   TIMEBOMB_ACTIVATED,
   EXPIRY_DATE,
-  EXPIRY_DAYS,
 } = require("../../variables/constants");
 
 const _verifyConfig = require("../verifyConfig");
@@ -30,13 +28,7 @@ try {
   const configData = config.getConfig();
 
   if (showVersion) {
-    console.log(
-      chalk.bold(
-        `${GLOBAL_NAME}, v${VERSION} (Build ${BUILD}${
-          RELEASE_CANDIDATE ? ` Release Candidate ${RELEASE_CANDIDATE}` : ""
-        })`
-      )
-    );
+    console.log(chalk.bold(`${GLOBAL_NAME}, v${VERSION} (Build ${BUILD})`));
   }
 
   // Only display if this is the first time BubbleOS is running
@@ -56,7 +48,7 @@ try {
         chalk.dim(
           `${chalk.bold(
             "WARNING!"
-          )} This beta software has a timebomb of +${EXPIRY_DAYS} days (since compile date).\nIt will expire on ${String(
+          )} This beta software has a timebomb of +90 days (since compile date).\nIt will expire on ${String(
             EXPIRY_DATE.getMonth() + 1
           ).padStart(2, "0")}/${String(EXPIRY_DATE.getDate()).padStart(
             2,