+++ src/variables/constants.js
@@ -16,32 +16,28 @@ const AUTHOR = "Arnav Thorat";
  *
  * Make sure this is a proper version number. Change this with the build (e.g. `78` will make the version `0.7.8`).
  *
- * For betas, add `-beta` to the end of the version number, and for release candidates, add `-rc<ver>`, replacing `<ver>` with the release candidate version.
+ * For betas, add `-beta` to the end of the version number.
  */
-const VERSION = "2.0.0-rc4";
+const VERSION = "2.0.0";
 /**
  * The build of the shell.
  *
  * Make sure this is a proper build number. Change this with the version (e.g. `0.7.8` will make the build `78`).
  */
 const BUILD = 200;
-/**
- * The release candidate of the shell. If set to `0`, it is not a release candidate.
- */
-const RELEASE_CANDIDATE = 4;
 
 /**
  * If the configuration file needs to be reset if it is an older version when loaded in this version.
  */
-const REQUIRE_CONFIG_RESET = false;
+const REQUIRE_CONFIG_RESET = true;
 /**
  * If the shell is in beta or not. `true` if it is in beta, else, `false`.
  */
-const IN_BETA = true;
+const IN_BETA = false;
 /**
  * If the timebomb for the shell has been activated. `true` if the timebomb is activated, else, `false`.
  */
-const TIMEBOMB_ACTIVATED = true;
+const TIMEBOMB_ACTIVATED = false;
 /**
  * The expiry date of the shell. If `TIMEBOMB_ACTIVATED` **and** `IN_BETA` are both `true`, the shell will always check on startup if the current date has surpassed the expiry date.
  *
@@ -60,14 +56,8 @@ const TIMEBOMB_ACTIVATED = true;
  * The expiry date is 90 days after the executable is compiled. This process is not automatic, so you must update this variable every time a new executable is compiled.
  *
  * To see the date in **90** days, [click here](https://www.google.com/search?q=Date+90+days+from+today).
- *
- * For release candidates, the expiry date is [30 days](https://www.google.com/search?q=Date+30+days+from+today) after the executable is compiled.
  */
-const EXPIRY_DATE = new Date(2025, 3, 9);
-/**
- * The number of days the expiry date is from the current date at the time of compilation.
- */
-const EXPIRY_DAYS = RELEASE_CANDIDATE ? 30 : 90;
+const EXPIRY_DATE = new Date(2025, 5, 20);
 
 module.exports = {
   GLOBAL_NAME,
@@ -75,10 +65,8 @@ module.exports = {
   AUTHOR,
   VERSION,
   BUILD,
-  RELEASE_CANDIDATE,
   REQUIRE_CONFIG_RESET,
   IN_BETA,
   TIMEBOMB_ACTIVATED,
   EXPIRY_DATE,
-  EXPIRY_DAYS,
 };